Entity collision fixes  (#9)
* Enemy collision fix.

* Entity logic move.

+++ deadline/src/main/java/org/game/message/EnemyMoveMessage.java
@@ -1,4 +1,4 @@
 package org.game.message;
 
-public record EnemyMoveMessage(int enemyId, int newX, int newY) implements Message {
+public record EnemyMoveMessage(long enemyId, int newX, int newY) implements Message {
 }
