+++ deadline/src/main/java/org/game/message/EnemySpawnMessage.java
@@ -3,5 +3,5 @@ package org.game.message;
 import org.game.entity.EnemySize;
 import org.game.entity.EnemyType;
 
-public record EnemySpawnMessage(int enemyId, EnemyType enemyType, EnemySize size, int initialX, int initialY) implements Message {
+public record EnemySpawnMessage(long enemyId, EnemyType enemyType, EnemySize size, int initialX, int initialY) implements Message {
 }
